src/requirements.txt
pyyaml==6.0.2
boto3==1.26.118
botocore==1.29.118
tqdm==4.66.5
numpy==1.24.3
pandas==1.5.3
PyWavelets==1.7.0
scikit-learn==1.3.2
scikit-image==0.24.0